Deduplicate test helpers
