+++ aya/src/maps/hash_map/per_cpu_hash_map.rs
@@ -150,10 +150,10 @@ impl<T: Borrow<MapData>, K: Pod, V: Pod> IterableMap<K, PerCpuValues<V>>
 
 #[cfg(test)]
 mod tests {
-    use super::{super::test_utils, *};
+    use super::*;
     use crate::{
         generated::bpf_map_type::{BPF_MAP_TYPE_LRU_PERCPU_HASH, BPF_MAP_TYPE_PERCPU_HASH},
-        maps::Map,
+        maps::{test_utils, Map},
     };
 
     #[test]
